Continue refactor to services
