+++ WMMAPI/Services/BaseService.cs
@@ -3,15 +3,15 @@ using System;
 using WMMAPI.Database;
 using WMMAPI.Interfaces;
 
-namespace WMMAPI.Repositories
+namespace WMMAPI.Services
 {
-    public abstract class BaseRepository<TEntity> where TEntity : class
+    public abstract class BaseService<TEntity> where TEntity : class
     {
         // Properties
         protected WMMContext Context { get; set; }
 
         // Base Constructor
-        public BaseRepository(WMMContext context)
+        public BaseService(WMMContext context)
         {
             Context = context;
         }
